requirements.txt
streamlit
numpy==1.26.4

langchain
langchain-core
langchain-community
langchain-text-splitters
faiss-cpu
pypdf


langchain-groq
langchain-huggingface

sentence-transformers

langgraph
langgraph-checkpoint
langgraph-checkpoint-postgres


psycopg[binary]
psycopg-pool


python-dotenv
pydantic

